+++ src/store/src/Bridge/Cache/Store.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\Store\Bridge\Local;
+namespace Symfony\AI\Store\Bridge\Cache;
 
 use Psr\Cache\CacheItemPoolInterface;
 use Symfony\AI\Platform\Vector\Vector;
@@ -26,7 +26,7 @@ use Symfony\Contracts\Cache\CacheInterface;
 /**
  * @author Guillaume Loulier <personal@guillaumeloulier.fr>
  */
-final class CacheStore implements ManagedStoreInterface, StoreInterface
+final class Store implements ManagedStoreInterface, StoreInterface
 {
     public function __construct(
         private readonly CacheInterface&CacheItemPoolInterface $cache,
@@ -34,7 +34,7 @@ final class CacheStore implements ManagedStoreInterface, StoreInterface
         private readonly string $cacheKey = '_vectors',
     ) {
         if (!interface_exists(CacheInterface::class)) {
-            throw new RuntimeException('For using the CacheStore as vector store, a symfony/contracts cache implementation is required. Try running "composer require symfony/cache" or another symfony/contracts compatible cache.');
+            throw new RuntimeException('For using the Cache store as vector store, a symfony/contracts cache implementation is required. Try running "composer require symfony/cache" or another symfony/contracts compatible cache.');
         }
     }
